+++ src/yafowil/loader.py
@@ -2,6 +2,7 @@
 import yafowil.utils
 
 
+@yafowil.utils.entry_point(order=0)
 def register():
     import yafowil.persistence
     import yafowil.common
@@ -11,11 +12,11 @@ def register():
 
 # execute all register entry points. supposed to be used for widget and theme
 # registration
-for ep in yafowil.utils.get_entry_points('register'):
-    ep.load()()
+for func in yafowil.utils.get_entry_points('register'):
+    func()
 
 
 # execute all configure entry points. supposed to be used for theme
 # configuration, like setting factory defaults and defining macros.
-for ep in yafowil.utils.get_entry_points('configure'):
-    ep.load()()                                              #pragma NO COVER
+for func in yafowil.utils.get_entry_points('configure'):
+    func()                                                    #pragma NO COVER